change inheritance
