--- app/controllers/accounts_controller.rb
@@ -1,4 +1,0 @@
-# frozen_string_literal: true
-
-class AccountsController < ApplicationController
-end
