+++ app/controllers/user_page/accounts_controller.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+module UserPage
+  class AccountsController < UserPage::ApplicationController
+  end
+end
