--- app/views/channels/index.json.jbuilder
@@ -1,3 +1,0 @@
-# frozen_string_literal: true
-
-json.array! @channels, partial: "channels/channel", as: :channel
