--- app/views/enterprises/index.json.jbuilder
@@ -1,3 +1,0 @@
-# frozen_string_literal: true
-
-json.array! @enterprises, partial: "enterprises/enterprise", as: :enterprise
