--- app/views/enterprises/show.json.jbuilder
@@ -1,3 +1,0 @@
-# frozen_string_literal: true
-
-json.partial! "enterprises/enterprise", enterprise: @enterprise
