--- app/views/messages/index.json.jbuilder
@@ -1,3 +1,0 @@
-# frozen_string_literal: true
-
-json.array! @messages, partial: "messages/message", as: :message
