--- app/views/messages/show.json.jbuilder
@@ -1,3 +1,0 @@
-# frozen_string_literal: true
-
-json.partial! "messages/message", message: @message
