+++ app/views/user_page/accounts/_account.json.jbuilder
