+++ app/views/user_page/accounts/index.json.jbuilder
