+++ app/views/user_page/accounts/show.json.jbuilder
