+++ app/views/user_page/channels/_channel.json.jbuilder
