+++ app/views/user_page/channels/index.json.jbuilder
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+json.array! @channels, partial: "channel", as: :channel
