+++ app/views/user_page/channels/show.json.jbuilder
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+json.partial! "channel", channel: @channel
