+++ app/views/user_page/enterprises/_enterprise.json.jbuilder
