+++ app/views/user_page/enterprises/index.json.jbuilder
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+json.array! @enterprises, partial: "enterprise", as: :enterprise
