+++ app/views/user_page/enterprises/show.json.jbuilder
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+json.partial! "enterprise", enterprise: @enterprise
