+++ app/views/user_page/messages/_message.json.jbuilder
