+++ app/views/user_page/messages/index.json.jbuilder
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+json.array! @messages, partial: "message", as: :message
